1. Replacing sync and async with simple and simple_async 2. Made mpi4py flags match
