+++ exarl/config/workflow_cfg/rma.json
@@ -1,6 +1,5 @@
 {
     "learner_procs": 1,
-    "process_per_env": 1,
     "data_structure": "buff_unchecked",
     "data_structure_length": 32,
     "max_model_lag": "none",
