--- exarl/config/workflow_cfg/simple.json
@@ -1,8 +1,0 @@
-{
-    "learner_procs": 1,
-    "episode_block": "false",
-    "batch_frequency": 1,
-    "save_weights_per_episode": "false",
-    "action_type": "variable",
-    "mpi4py_rc": "true"
-}
