+++ exarl/workflows/__init__.py
@@ -24,13 +24,3 @@ register(
     id='random',
     entry_point='exarl.workflows.workflow_vault:RANDOM'
 )
-
-register(
-    id='simple',
-    entry_point='exarl.workflows.workflow_vault:SIMPLE'
-)
-
-register(
-    id='simple_async',
-    entry_point='exarl.workflows.workflow_vault:SIMPLE_ASYNC'
-)
