+++ exarl/workflows/workflow_vault/__init__.py
@@ -2,5 +2,3 @@ from exarl.workflows.workflow_vault.sync_learner import SYNC
 from exarl.workflows.workflow_vault.async_learner import ASYNC
 from exarl.workflows.workflow_vault.rma_learner import RMA
 from exarl.workflows.workflow_vault.random_learner import RANDOM
-from exarl.workflows.workflow_vault.simple_learner import SIMPLE
-from exarl.workflows.workflow_vault.simple_async_learner import SIMPLE_ASYNC
